move ModelServer tests to test_package and delete stateless module files
